# swagger/swagger.yaml
consumes:
- application/json
- application/xml
- application/gob
- application/x-gob
definitions:
  AddItemPayload:
    example:
      item: Nobis tempore eum.
      member_price: 2.494102e+16
      price: 1.29947e+18
      quantity: 5.5559257e+17
    properties:
      item:
        description: item name
        example: Nobis tempore eum.
        type: string
      member_price:
        description: Membership discount
        example: 2.494102e+16
        format: int64
        type: integer
      price:
        description: item price
        example: 1.29947e+18
        format: int64
        type: integer
      quantity:
        description: item quantity
        example: 5.5559257e+17
        format: int64
        type: integer
    required:
    - item
    - price
    - member_price
    - quantity
    title: AddItemPayload
    type: object
  AddOrderPayload:
    example:
      date: 2016-01-18T00:05:35+09:00
      item_id: 5.152369e+18
      price: 8.146832e+18
      quantity: 6.6346896e+17
      register: Aut illo.
    properties:
      date:
        description: Order date
        example: 2016-01-18T00:05:35+09:00
        format: date-time
        type: string
      item_id:
        description: Unique item ID
        example: 5.152369e+18
        format: int64
        type: integer
      price:
        description: Item price
        example: 8.146832e+18
        format: int64
        type: integer
      quantity:
        description: Item quantity
        example: 6.6346896e+17
        format: int64
        type: integer
      register:
        description: Register's name
        example: Aut illo.
        type: string
    required:
    - item_id
    - quantity
    - price
    - register
    - date
    title: AddOrderPayload
    type: object
  GoaExampleOrders:
    description: An order (default view)
    example:
      available: true
      date: 1991-10-11T00:10:20+09:00
      id: 4.4999475e+17
      item: Ducimus quasi.
      item_id: 2.1326141e+18
      price: 7.137601e+18
      quantity: 3.1358608e+18
      register: Quia natus voluptates maiores nemo quia.
    properties:
      available:
        description: Is available order
        example: true
        type: boolean
      date:
        description: Order date
        example: 1991-10-11T00:10:20+09:00
        format: date-time
        type: string
      id:
        description: Unique order ID
        example: 4.4999475e+17
        format: int64
        type: integer
      item:
        description: Item name
        example: Ducimus quasi.
        type: string
      item_id:
        description: Unique item ID
        example: 2.1326141e+18
        format: int64
        type: integer
      price:
        description: Item price
        example: 7.137601e+18
        format: int64
        type: integer
      quantity:
        description: Item quantity
        example: 3.1358608e+18
        format: int64
        type: integer
      register:
        description: Register's name
        example: Quia natus voluptates maiores nemo quia.
        type: string
    required:
    - id
    - item_id
    - item
    - quantity
    - price
    - register
    - date
    - available
    title: 'Mediatype identifier: application/vnd.goa.example.orders+json; view=default'
    type: object
  GoaExampleOrdersCollection:
    description: GoaExampleOrdersCollection is the media type for an array of GoaExampleOrders
      (default view)
    example:
    - available: true
      date: 1991-10-11T00:10:20+09:00
      id: 4.4999475e+17
      item: Ducimus quasi.
      item_id: 2.1326141e+18
      price: 7.137601e+18
      quantity: 3.1358608e+18
      register: Quia natus voluptates maiores nemo quia.
    items:
      $ref: '#/definitions/GoaExampleOrders'
    title: 'Mediatype identifier: application/vnd.goa.example.orders+json; type=collection;
      view=default'
    type: array
  GoaExampleRegisysGoods:
    description: An Item (default view)
    example:
      id: 2.9891975e+18
      item: Quia consequuntur.
      member_price: 5.148723e+18
      price: 8.0569887e+18
      quantity: 8.9942954e+18
      user_id: 5.9552595e+18
      user_name: Dolorem consequatur optio delectus.
    properties:
      id:
        description: Unique item ID
        example: 2.9891975e+18
        format: int64
        type: integer
      item:
        description: item name
        example: Quia consequuntur.
        type: string
      member_price:
        description: Membership discount
        example: 5.148723e+18
        format: int64
        type: integer
      price:
        description: item price
        example: 8.0569887e+18
        format: int64
        type: integer
      quantity:
        description: item quantity
        example: 8.9942954e+18
        format: int64
        type: integer
      user_id:
        description: Unique User ID
        example: 5.9552595e+18
        format: int64
        type: integer
      user_name:
        description: Username
        example: Dolorem consequatur optio delectus.
        type: string
    required:
    - id
    - item
    - price
    - member_price
    - user_id
    - user_name
    - quantity
    title: 'Mediatype identifier: application/vnd.goa.example.regisys.goods+json;
      view=default'
    type: object
  GoaExampleRegisysGoodsCollection:
    description: GoaExampleRegisysGoodsCollection is the media type for an array of
      GoaExampleRegisysGoods (default view)
    example:
    - id: 2.9891975e+18
      item: Quia consequuntur.
      member_price: 5.148723e+18
      price: 8.0569887e+18
      quantity: 8.9942954e+18
      user_id: 5.9552595e+18
      user_name: Dolorem consequatur optio delectus.
    items:
      $ref: '#/definitions/GoaExampleRegisysGoods'
    title: 'Mediatype identifier: application/vnd.goa.example.regisys.goods+json;
      type=collection; view=default'
    type: array
  GoaExampleToken:
    description: Username and ID (default view)
    example:
      id: 8.5014976e+18
      username: Quis hic facere distinctio.
    properties:
      id:
        description: Unique user ID
        example: 8.5014976e+18
        format: int64
        type: integer
      username:
        description: Username
        example: Quis hic facere distinctio.
        type: string
    required:
    - username
    - id
    title: 'Mediatype identifier: application/vnd.goa.example.token+json; view=default'
    type: object
  GoaExampleUser:
    description: Users (default view)
    example:
      group: Reprehenderit iusto voluptatibus sed.
      id: 4.989876e+18
      name: Inventore sit unde.
    properties:
      group:
        description: Group of user
        example: Reprehenderit iusto voluptatibus sed.
        type: string
      id:
        description: Unique user ID
        example: 4.989876e+18
        format: int64
        type: integer
      name:
        description: Username
        example: Inventore sit unde.
        type: string
    required:
    - id
    - name
    - group
    title: 'Mediatype identifier: application/vnd.goa.example.user+json; view=default'
    type: object
  GoaExampleUserCollection:
    description: GoaExampleUserCollection is the media type for an array of GoaExampleUser
      (default view)
    example:
    - group: Reprehenderit iusto voluptatibus sed.
      id: 4.989876e+18
      name: Inventore sit unde.
    items:
      $ref: '#/definitions/GoaExampleUser'
    title: 'Mediatype identifier: application/vnd.goa.example.user+json; type=collection;
      view=default'
    type: array
  ModifyItemPayload:
    example:
      id: 8.2702076e+18
      item: Optio molestiae magni quae ut tenetur.
      member_price: 5.203682e+18
      price: 5.2072132e+17
      quantity: 6.3114837e+18
    properties:
      id:
        description: Unique item ID
        example: 8.2702076e+18
        format: int64
        type: integer
      item:
        description: item name
        example: Optio molestiae magni quae ut tenetur.
        type: string
      member_price:
        description: Membership discount
        example: 5.203682e+18
        format: int64
        type: integer
      price:
        description: item price
        example: 5.2072132e+17
        format: int64
        type: integer
      quantity:
        description: item quantity
        example: 6.3114837e+18
        format: int64
        type: integer
    required:
    - id
    title: ModifyItemPayload
    type: object
host: localhost:8080
info:
  title: Register System for Junk-Ichi in MMA
  version: ""
paths:
  /goods:
    get:
      description: |-
        Get items

        Required security scopes:
          * `api:access`
      operationId: goods#show
      parameters:
      - description: User ID
        in: query
        name: user
        required: false
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/GoaExampleRegisysGoodsCollection'
        "404":
          description: Not Found
      schemes:
      - http
      security:
      - jwt:
        - api:access
      summary: show goods
      tags:
      - goods
    post:
      description: |-
        Add items

        Required security scopes:
          * `api:access`
      operationId: goods#add
      parameters:
      - in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/AddItemPayload'
      responses:
        "204":
          description: No Content
      schemes:
      - http
      security:
      - jwt:
        - api:access
      summary: add goods
      tags:
      - goods
  /goods/{id}:
    delete:
      description: |-
        Delete item

        Required security scopes:
          * `api:access`
      operationId: goods#delete
      parameters:
      - description: Unique item ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "403":
          description: Forbidden
        "404":
          description: Not Found
      schemes:
      - http
      security:
      - jwt:
        - api:access
      summary: delete goods
      tags:
      - goods
    put:
      description: |-
        Modify item

        Required security scopes:
          * `api:access`
      operationId: goods#modify
      parameters:
      - description: Unique item ID
        in: path
        name: id
        required: true
        type: integer
      - in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/ModifyItemPayload'
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
      schemes:
      - http
      security:
      - jwt:
        - api:access
      summary: modify goods
      tags:
      - goods
  /orders:
    get:
      description: |-
        Get orders

        Required security scopes:
          * `api:access`
      operationId: orders#show
      parameters:
      - description: Order date
        in: query
        name: date
        required: false
        type: string
      - description: Unique user ID
        in: query
        name: user
        required: false
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/GoaExampleOrdersCollection'
        "404":
          description: Not Found
      schemes:
      - http
      security:
      - jwt:
        - api:access
      summary: show orders
      tags:
      - orders
    post:
      description: |-
        Add order

        Required security scopes:
          * `api:access`
      operationId: orders#add
      parameters:
      - in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/AddOrderPayload'
      responses:
        "204":
          description: No Content
        "403":
          description: Forbidden
      schemes:
      - http
      security:
      - jwt:
        - api:access
      summary: add orders
      tags:
      - orders
  /orders/{id}:
    delete:
      description: |-
        Disable order

        Required security scopes:
          * `api:access`
      operationId: orders#delete
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "403":
          description: Forbidden
        "404":
          description: Not Found
      schemes:
      - http
      security:
      - jwt:
        - api:access
      summary: delete orders
      tags:
      - orders
  /token:
    get:
      description: Create a valid JWT
      operationId: jwt#signin
      parameters:
      - description: Is member of MMA
        in: query
        name: is_member
        required: true
        type: boolean
      responses:
        "200":
          description: OK
          headers:
            Authorization:
              description: Generated JWT
              type: string
          schema:
            $ref: '#/definitions/GoaExampleToken'
        "401":
          description: Unauthorized
      schemes:
      - http
      security:
      - SigninBasicAuth: []
      summary: signin jwt
      tags:
      - jwt
  /user:
    post:
      description: Add user for NOT MMA member)
      operationId: user#add
      responses:
        "204":
          description: No Content
        "403":
          description: Forbidden
      schemes:
      - http
      security:
      - SigninBasicAuth: []
      summary: add user
      tags:
      - user
  /user/{id}:
    get:
      description: |-
        Show one user

        Required security scopes:
          * `api:access`
      operationId: user#show
      parameters:
      - description: Unique user ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/GoaExampleUser'
        "404":
          description: Not Found
      schemes:
      - http
      security:
      - jwt:
        - api:access
      summary: show user
      tags:
      - user
    put:
      description: |-
        Modify group of user

        Required security scopes:
          * `api:access`
      operationId: user#modify
      parameters:
      - description: Group of user
        enum:
        - admin
        - register
        - none
        in: query
        name: group
        required: false
        type: string
      - in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "403":
          description: Forbidden
        "404":
          description: Not Found
      schemes:
      - http
      security:
      - jwt:
        - api:access
      summary: modify user
      tags:
      - user
  /user/list:
    get:
      description: |-
        Show users list

        Required security scopes:
          * `api:access`
      operationId: user#showList
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/GoaExampleUserCollection'
      schemes:
      - http
      security:
      - jwt:
        - api:access
      summary: showList user
      tags:
      - user
produces:
- application/json
- application/xml
- application/gob
- application/x-gob
responses:
  Forbidden:
    description: Forbidden
  NoContent:
    description: No Content
  NotFound:
    description: Not Found
  Unauthorized:
    description: Unauthorized
schemes:
- http
securityDefinitions:
  SigninBasicAuth:
    type: basic
  jwt:
    description: |2-


      **Security Scopes**:
        * `api:access`: API access
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"